+++ tests/admin.py
@@ -2,9 +2,14 @@ from django.contrib import admin
 
 from conman.routes.admin import RouteChildAdmin
 
-from .models import NestedRouteSubclass, RouteSubclass, URLConfRoute
+from .models import (
+    NestedRouteSubclass,
+    RouteSubclass,
+    TemplateRoute,
+    URLConfRoute,
+)
 
 
-models = (NestedRouteSubclass, RouteSubclass, URLConfRoute)
+models = (NestedRouteSubclass, RouteSubclass, TemplateRoute, URLConfRoute)
 for model in models:
     admin.site.register(model, RouteChildAdmin)