+++ tests/models.py
@@ -1,4 +1,6 @@
-from conman.routes.handlers import URLConfHandler
+from django.db import models
+
+from conman.routes.handlers import TemplateHandler, URLConfHandler
 from conman.routes.managers import RouteManager
 from conman.routes.models import Route
 
@@ -16,6 +18,15 @@ class RouteSubclass(Route):
 
 class NestedRouteSubclass(RouteSubclass):
     """A Route for testing nested subclasses of Route."""
+    # Silence RemovedInDjango20Warning about manager inheritance.
+    objects = RouteManager()
+
+
+class TemplateRoute(Route):
+    """A Route for testing TemplateHandler."""
+    content = models.TextField()
+    handler_class = TemplateHandler
+    template_name = 'basic_template.html'
     # Silence RemovedInDjango20Warning about manager inheritance.
     objects = RouteManager()